Second
2

+++ PersonalAssistant/Contacts/views.py
@@ -3,27 +3,17 @@ from .forms import ContactForm
 from django.shortcuts import render
 from datetime import date, timedelta
 from .models import Contact
-
-def add_contact(request):
-    if request.method == 'POST':
-        form = ContactForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return redirect('contact_list')  
-    else:
-        form = ContactForm()
-    return render(request, 'contacts/add_contact.html', {'form': form})
+from .forms import ContactDeleteForm
 
 
 def start_page(request):
     context = {}
     return render(request, 'Contacts/base.html', context)
 
-
 def card_subtitle_view(request):
     return render(request, 'Contacts/card_subtitle.html')
 
-
+#добавление контактов 
 def add_contact(request):
     if request.method == 'POST':
         form = ContactForm(request.POST)
@@ -33,29 +23,6 @@ def add_contact(request):
     else:
         form = ContactForm()
     return render(request, 'contacts/add_contact.html', {'form': form})
-
-
-
-# поиск контактов, у которых день рождения через заданное количество дней от текущей даты
-def upcoming_birthdays(request, days=7):
-    # Определение текущей даты
-    today = date.today()
-
-    # Определение дату через заданное количество дней
-    future_date = today + timedelta(days=days)
-
-    # Определение контактов, у которых дни рождения между текущей датой и будущей датой
-    upcoming_contacts = Contact.objects.filter(
-        birth_date__gte=today, birth_date__lte=future_date
-    )
-
-    context = {
-        'upcoming_contacts': upcoming_contacts,
-        'future_date': future_date,
-    }
-
-    return render(request, 'contacts/upcoming_birthdays.html', context)
-
 
 
 #поиск контактов среди контактов книги
@@ -72,8 +39,8 @@ def contact_search(request):
         'query': query,
         'results': results,
     }
-
     return render(request, 'contacts/contact_search.html', context)
+
 
 # редактирование контактов
 def edit_contact(request, contact_id):
@@ -90,12 +57,42 @@ def edit_contact(request, contact_id):
     return render(request, 'contacts/edit_contact.html', {'form': form})
 
 
-# удаление контактов
 def delete_contact(request, contact_id):
     contact = get_object_or_404(Contact, id=contact_id)
 
     if request.method == 'POST':
-        contact.delete()
-        return redirect('contact_list')  # Перенаправление на страницу со списком контактов
+        form = ContactDeleteForm(request.POST)
+        if form.is_valid():
+            contact.delete()
+            return redirect('contact_list')  # Перенаправление на страницу со списком контактов
+    else:
+        form = ContactDeleteForm()
 
-    return render(request, 'contacts/delete_contact.html', {'contact': contact})
+    return render(request, 'contacts/delete_contact.html', {'contact': contact, 'form': form})
+
+
+# список всех контактов из базы данных
+def contact_list(request):
+    
+    contacts = Contact.objects.all()
+
+    context = {
+        'contacts': contacts,
+    }
+
+    return render(request, 'contacts/contact_list.html', context)
+
+
+
+def calculate_days_until_birthday(birthday):
+    today = date.today()
+    
+    # Проверяем, есть ли день рождения в этом году или следующем году
+    next_birthday = birthday.replace(year=today.year)
+    if next_birthday < today:
+        next_birthday = next_birthday.replace(year=today.year + 1)
+    
+    # Рассчитываем количество дней до дня рождения
+    days_until_birthday = (next_birthday - today).days
+    
+    return days_until_birthday
